Syscall rewriter (#42)
This PR adds the support of static binary rewriting that hooks syscalls
in an ELF file. We need this functionality because we would like to
redirect syscalls made by an app to litebox when the latter runs in the
user mode. We modify each syscall to add an indirect jmp to a trampoline
. All trampolines are stored in a new .trampoline section. The
trampoline section starts with 8 bytes which will be used by the litebox
loader to store the syscall entry in litebox.

---------

Co-authored-by: Jay Bosamiya (Microsoft) <jayb@microsoft.com>