Syscall handler updates (#157)
While this design is not (yet) in the full ideal state I have in mind,
this should cover the biggest downside (imho) to the existing design for
connecting the handler between the `platform` and the `shim` crates.
Specifically, rather than being implicit via the linker (which is a very
C-like way to handle the issue), this uses an explicit registration of a
Rust function. _Technically_, this costs us an `RwLock` read, but that
is _very_ cheap for something that is never updated except at boot
(thus, there is never any contention on that lock, even with multiple
readers on it).

There is some more refactoring and cleanup that I have done along the
way (including adding an explicit option to our runner for seccomp or
not).

In terms of the movement of code, a non-trivial change during the first
commit in this PR (ac2ebfc839c02cc9e8263df097a85cfb155bced7) rather than
just moving code over is that I moved the handlers for
`Exit`/`ExitProvider`/`Ret` to the actual handling, rather than
happening implicitly while doing the dispatch. Additionally, it seemed
to fit in better to define this translation right beside the definition
of the request enum itself, thus it is sitting in common now.