Improved fs/net metadata handling (#370)
This PR cleans up metadata handling by migrating the handling from the
net subsystem, and each fs that we support, into the fd subsystem. This
reduces the amount of redundant code we have and makes later migration
in the shim easier.

This PR is primarily semantics-preserving refactoring, but _does_ change
our north interfaces a bit. Specifically, we no longer will have the
parametric metadata directly accessible at the fs/net north interfaces,
but instead are linked to the fd interface. I've fixed up the relevant
calls in the shim.

Interestingly, the net subsystem's metadata support was lacking behind
the fs subsystem's metadata support (network FDs could previously not
store metadata separately at the FD level and at entry level). With this
PR, it is up to full parity in terms of support.

---

This is yet another PR in the series of PRs towards the better FD design
(connected to #31).