Raw FDs: free slots + explicit insertion (#373)
This PR updates the LiteBox raw FD management. Specifically, it makes
two changes:

1. Uses free slots for raw FD storage, thereby ensuring they are smaller
if you are picking a slot that without explicitly asking for one. This
improves over the previous design which used an internal numbering that
would still be small numbers, but might be more surprising behavior
(since the internal numbers tick up faster due to the existence of the
layered FS).
2. introduces a new API that lets you request a specific raw integer FD
slot. This is useful for the stdin/stdout/stderr FDs, since we want them
to be 0,1,2 on some shims (Linux). It is also useful to support nicer
`dup2` in a future PR.

---

This is yet another PR in the series of PRs towards the better FD design
(connected to https://github.com/microsoft/litebox/issues/31).