Refactor ratchet function for better granularity (#401)
This PR is purely to make it easier to notice at a quick glance which
particular components are contributing to the ratchets, and whether we
need to focus on reducing them in any particular one, or it is ok to
keep them as-is.

No functionality change, just a refactoring to make the ratchets easier
to keep track of.