Handle execve fully within the shim (#426)
Instead of calling out to the platform to load and run the next program,
handle the execve syscall mostly within the Linux shim (a callback has
been preserved to allow the Linux userland shim to update auxv and
envp.)

This requires fixing the platforms to properly restore rsp and rip if
they are changed in the context structure as part of the syscall. This
has been done for the 64-bit userland platforms.

Note that execve still does not work on Windows userland, apparently due
to some failure when unmapping memory.