Add an integration test for python (#449)
Add a simple end-to-end test for python3.

As I commented in the code, pre-compiled .pyc files are not used because
of the lack of support for timestamp in fs.

Also, updated the function `run_target_program` so that it can 1) enable
network (i.e., tun), 2) pass environment variables, and 3) create tar
file using `ustar` format.