Fix setting new TLS (#470)
Finally figure out the root cause of the issue I mentioned in the
[comment](https://github.com/microsoft/litebox/pull/463#issuecomment-3487385084),
which is caused by incorrect handling of TLS when calling `sys_clone`
with TLS (`*mut UserDesc`). The problem was that we set TLS for the new
thread in the new thread's context instead of its parent's (in Linux,
kernel can set TLS for another thread while user code cannot), which
means after `sys_clone` returns and before the child runs, the parent
may reset `*mut UserDesc`, resulting in the child setting the wrong TLS.
This only affects x86, because on x64 we only need to set `fsbase` with
the raw pointer.

To address it, we save a copy of `UserDesc` instead of the original
pointer passed from user space. One caveat is that if we need to
allocate a new entry for TLS (i.e., `UserDesc.entry_number == -1` for
the new thread, we don't write the new entry number back to user space
because the original pointer to `UserDesc` might be invalid as mentioned
above. This shouldn't be a problem in practice. The entry number for a
new thread is usually the same as the main thread, so no need for
allocating a new one.